iox-#1196 Refine some integer types (signed/unsigned)

+++ iceoryx_hoofs/test/moduletests/test_cxx_algorithm.cpp
@@ -123,16 +123,16 @@ TEST_F(algorithm_test, DoesContainValue_ValueListOfMultipleValuesDoesContainValu
 TEST_F(algorithm_test, MergeTwoDisjunctNonEmptySortedContainers)
 {
     ::testing::Test::RecordProperty("TEST_ID", "4f39641f-de8a-434a-8a50-cd2b66b476da");
-    constexpr int64_t OFFSET = 1337;
-    vector<int64_t, 10U> first;
-    vector<int64_t, 10U> second;
+    constexpr uint64_t OFFSET = 1337U;
+    vector<uint64_t, 10U> first;
+    vector<uint64_t, 10U> second;
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         first.emplace_back(i + OFFSET);
     }
 
-    for (int64_t i = 5; i < 10; ++i)
+    for (uint64_t i = 5U; i < 10U; ++i)
     {
         second.emplace_back(i + OFFSET);
     }
@@ -141,9 +141,9 @@ TEST_F(algorithm_test, MergeTwoDisjunctNonEmptySortedContainers)
     auto mergedContainerSwitched = uniqueMergeSortedContainers(second, first);
 
     ASSERT_THAT(mergedContainer.size(), Eq(10U));
-    for (int64_t i = 0; i < 10; ++i)
+    for (uint64_t i = 0U; i < 10U; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i + OFFSET));
+        EXPECT_THAT(mergedContainer[i], Eq(i + OFFSET));
     }
     EXPECT_TRUE(mergedContainer == mergedContainerSwitched);
 }
@@ -151,17 +151,17 @@ TEST_F(algorithm_test, MergeTwoDisjunctNonEmptySortedContainers)
 TEST_F(algorithm_test, MergeTwoDisjunctNonEmptySortedContainersWithAGap)
 {
     ::testing::Test::RecordProperty("TEST_ID", "15d3c063-8bc5-47eb-84a4-35f055a1d82c");
-    constexpr int64_t OFFSET = 41;
-    constexpr int64_t GAP = 13;
-    vector<int64_t, 10U> first;
-    vector<int64_t, 10U> second;
+    constexpr uint64_t OFFSET = 41U;
+    constexpr uint64_t GAP = 13;
+    vector<uint64_t, 10U> first;
+    vector<uint64_t, 10U> second;
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         first.emplace_back(i + OFFSET);
     }
 
-    for (int64_t i = 5; i < 10; ++i)
+    for (uint64_t i = 5U; i < 10U; ++i)
     {
         second.emplace_back(i + OFFSET + GAP);
     }
@@ -170,13 +170,13 @@ TEST_F(algorithm_test, MergeTwoDisjunctNonEmptySortedContainersWithAGap)
     auto mergedContainerSwitched = uniqueMergeSortedContainers(second, first);
 
     ASSERT_THAT(mergedContainer.size(), Eq(10U));
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i + OFFSET));
+        EXPECT_THAT(mergedContainer[i], Eq(i + OFFSET));
     }
-    for (int64_t i = 5; i < 10; ++i)
+    for (uint64_t i = 5U; i < 10U; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i + OFFSET + GAP));
+        EXPECT_THAT(mergedContainer[i], Eq(i + OFFSET + GAP));
     }
     EXPECT_TRUE(mergedContainer == mergedContainerSwitched);
 }
@@ -184,16 +184,16 @@ TEST_F(algorithm_test, MergeTwoDisjunctNonEmptySortedContainersWithAGap)
 TEST_F(algorithm_test, MergeTwoAlternatingDisjunctNonEmptySortedContainers)
 {
     ::testing::Test::RecordProperty("TEST_ID", "02cc9514-6cfe-4e08-8806-f371561fef41");
-    constexpr int64_t OFFSET = 4301;
-    vector<int64_t, 10U> first;
-    vector<int64_t, 10U> second;
+    constexpr uint64_t OFFSET = 4301U;
+    vector<uint64_t, 10U> first;
+    vector<uint64_t, 10U> second;
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         first.emplace_back(i * 2 + OFFSET);
     }
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         second.emplace_back(i * 2 + 1 + OFFSET);
     }
@@ -202,9 +202,9 @@ TEST_F(algorithm_test, MergeTwoAlternatingDisjunctNonEmptySortedContainers)
     auto mergedContainerSwitched = uniqueMergeSortedContainers(second, first);
 
     ASSERT_THAT(mergedContainer.size(), Eq(10U));
-    for (int64_t i = 0; i < 10; ++i)
+    for (uint64_t i = 0; i < 10; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i + OFFSET));
+        EXPECT_THAT(mergedContainer[i], Eq(i + OFFSET));
     }
     EXPECT_TRUE(mergedContainer == mergedContainerSwitched);
 }
@@ -212,10 +212,10 @@ TEST_F(algorithm_test, MergeTwoAlternatingDisjunctNonEmptySortedContainers)
 TEST_F(algorithm_test, MergingIdenticalContainerResultsInUnchangedContainer)
 {
     ::testing::Test::RecordProperty("TEST_ID", "50f05cf2-62fa-49b8-8380-1dd0ac2470ec");
-    constexpr int64_t OFFSET = 313;
-    vector<int64_t, 10U> someContainer;
+    constexpr uint64_t OFFSET = 313U;
+    vector<uint64_t, 10U> someContainer;
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         someContainer.emplace_back(i * 2 + OFFSET);
     }
@@ -223,7 +223,7 @@ TEST_F(algorithm_test, MergingIdenticalContainerResultsInUnchangedContainer)
     auto mergedContainer = uniqueMergeSortedContainers(someContainer, someContainer);
 
     ASSERT_THAT(mergedContainer.size(), Eq(5U));
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i * 2 + OFFSET));
     }
@@ -232,36 +232,36 @@ TEST_F(algorithm_test, MergingIdenticalContainerResultsInUnchangedContainer)
 TEST_F(algorithm_test, MergingWithOneEmptyContainerResultsInUnchangedContainer)
 {
     ::testing::Test::RecordProperty("TEST_ID", "b0a0eb3a-08a3-4898-a8c9-a4f7eff0115c");
-    constexpr int64_t OFFSET = 707;
-    vector<int64_t, 10U> someContainer;
+    constexpr uint64_t OFFSET = 707U;
+    vector<uint64_t, 10U> someContainer;
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         someContainer.emplace_back(i * 3 + OFFSET);
     }
 
-    auto mergedContainer = uniqueMergeSortedContainers(someContainer, vector<int64_t, 10U>());
+    auto mergedContainer = uniqueMergeSortedContainers(someContainer, vector<uint64_t, 10U>());
 
     ASSERT_THAT(mergedContainer.size(), Eq(5U));
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i * 3 + OFFSET));
+        EXPECT_THAT(mergedContainer[i], Eq(i * 3 + OFFSET));
     }
 }
 
 TEST_F(algorithm_test, MergePartiallyOverlappingSortedContainers)
 {
     ::testing::Test::RecordProperty("TEST_ID", "c57dda77-81a5-413f-b54b-e924e67b66a5");
-    constexpr int64_t OFFSET = 8055;
-    vector<int64_t, 10U> first;
-    vector<int64_t, 10U> second;
+    constexpr uint64_t OFFSET = 8055U;
+    vector<uint64_t, 10U> first;
+    vector<uint64_t, 10U> second;
 
-    for (int64_t i = 3; i < 10; ++i)
+    for (uint64_t i = 3U; i < 10U; ++i)
     {
         first.emplace_back(i + OFFSET);
     }
 
-    for (int64_t i = 0; i < 8; ++i)
+    for (uint64_t i = 0U; i < 8U; ++i)
     {
         second.emplace_back(i + OFFSET);
     }
@@ -270,9 +270,9 @@ TEST_F(algorithm_test, MergePartiallyOverlappingSortedContainers)
     auto mergedContainerSwitched = uniqueMergeSortedContainers(second, first);
 
     ASSERT_THAT(mergedContainer.size(), Eq(10U));
-    for (int64_t i = 0; i < 10; ++i)
+    for (uint64_t i = 0U; i < 10U; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i + OFFSET));
+        EXPECT_THAT(mergedContainer[i], Eq(i + OFFSET));
     }
     EXPECT_TRUE(mergedContainer == mergedContainerSwitched);
 }
@@ -280,11 +280,11 @@ TEST_F(algorithm_test, MergePartiallyOverlappingSortedContainers)
 TEST_F(algorithm_test, MergeWithDisjunctOneElementContainer)
 {
     ::testing::Test::RecordProperty("TEST_ID", "7a56b0f9-82d2-4f9a-881f-338dd572a453");
-    constexpr int64_t OFFSET = 333331;
-    vector<int64_t, 10U> first;
-    vector<int64_t, 10U> second;
+    constexpr uint64_t OFFSET = 333331U;
+    vector<uint64_t, 10U> first;
+    vector<uint64_t, 10U> second;
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         first.emplace_back(i + OFFSET);
     }
@@ -295,9 +295,9 @@ TEST_F(algorithm_test, MergeWithDisjunctOneElementContainer)
     auto mergedContainerSwitched = uniqueMergeSortedContainers(second, first);
 
     ASSERT_THAT(mergedContainer.size(), Eq(6U));
-    for (int64_t i = 0; i < 6; ++i)
+    for (uint64_t i = 0U; i < 6U; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i + OFFSET));
+        EXPECT_THAT(mergedContainer[i], Eq(i + OFFSET));
     }
     EXPECT_TRUE(mergedContainer == mergedContainerSwitched);
 }
@@ -305,11 +305,11 @@ TEST_F(algorithm_test, MergeWithDisjunctOneElementContainer)
 TEST_F(algorithm_test, MergeWithOverlappingOneElementContainer)
 {
     ::testing::Test::RecordProperty("TEST_ID", "05fb7baf-51e9-4ff9-bb35-8ae4174b0216");
-    constexpr int64_t OFFSET = 29292929;
-    vector<int64_t, 10U> first;
-    vector<int64_t, 10U> second;
+    constexpr uint64_t OFFSET = 29292929U;
+    vector<uint64_t, 10U> first;
+    vector<uint64_t, 10U> second;
 
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
         first.emplace_back(i + OFFSET);
     }
@@ -320,9 +320,9 @@ TEST_F(algorithm_test, MergeWithOverlappingOneElementContainer)
     auto mergedContainerSwitched = uniqueMergeSortedContainers(second, first);
 
     ASSERT_THAT(mergedContainer.size(), Eq(5U));
-    for (int64_t i = 0; i < 5; ++i)
+    for (uint64_t i = 0U; i < 5U; ++i)
     {
-        EXPECT_THAT(mergedContainer[static_cast<uint64_t>(i)], Eq(i + OFFSET));
+        EXPECT_THAT(mergedContainer[i], Eq(i + OFFSET));
     }
     EXPECT_TRUE(mergedContainer == mergedContainerSwitched);
 }
